pyproject.toml
[tool.poetry] # This is the old, proprietary Poetry metadata section
name = "system-testing-demo-bdd"
version = "0.1.0"
description = ""
authors = ["Sayan Roy <sayan.roy@thoughtworks.com>"]
readme = "README.md"



[tool.poetry.dependencies] # Dependencies must be under this tool-specific section
python = ">=3.10,<3.11"
requests = {version = ">=2.32.5,<3.0.0"}
pytest = "*"
pytest-bdd = "*"
pytest-playwright = "*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"